// clients/accountingService/models/ShareClassCreateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ShareClassCreateDto = {
    readonly id?: string;
    readonly timestamp?: string;
    name?: string | null;
    value?: boolean;
    description?: string | null;
    forexRates?: string | null;
    currencyId?: string | null;
    tenantId?: string | null;
    enrolmentId?: string | null;
};