// clients/accountingService/models/ShareClassUpdateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ShareClassUpdateDto = {
    name?: string | null;
    value?: boolean;
    description?: string | null;
    forexRates?: string | null;
    currencyId?: string | null;
    tenantId?: string | null;
    enrolmentId?: string | null;
};