// clients/accountingService/models/ShareIssuanceCreateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ShareIssuanceCreateDto = {
    readonly id?: string;
    readonly timestamp?: string;
    tenantId?: string | null;
    enrolmentId?: string | null;
    unitPrice?: number;
    quantity?: number;
    currencyId?: string | null;
};