// clients/accountingService/models/ShareIssuanceUpdateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ShareIssuanceUpdateDto = {
    tenantId?: string | null;
    enrolmentId?: string | null;
    unitPrice?: number;
    quantity?: number;
    currencyId?: string | null;
};