// clients/accountingService/models/ShareTransferCreateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ShareTransferCreateDto = {
    readonly id?: string;
    readonly timestamp?: string;
    description?: string | null;
    value?: number;
    newShareHolderId?: string | null;
    formerShareHolderId?: string | null;
    shareTransferReasonId?: string | null;
    enrolmentId?: string | null;
    tenantId?: string | null;
};