// clients/accountingService/models/ShareTransferReasonCreateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ShareTransferReasonCreateDto = {
    readonly id?: string;
    readonly timestamp?: string;
    name?: string | null;
    description?: string | null;
    enrolmentId?: string | null;
    tenantId?: string | null;
};